+++ storybook/stories/settings/general/General.stories.tsx
@@ -37,6 +37,8 @@ const getParamName = (obj, itemName): any =>
   Object.entries(obj).find((entry: [any, any]) => itemName === entry[1]);
 
 /* eslint-disable consistent-return */
+// @ts-ignore
+// @ts-ignore
 storiesOf('Settings / General', module)
   .addDecorator(SettingsWrapper) // ====== Stories ======
   .add(
@@ -99,6 +101,7 @@ storiesOf('Settings / General', module)
       }}
     />
   ))
+  // @ts-ignore ts-migrate(2345)
   .add('Terms of Service', (_, props) => {
     const termsOfUseSource = require(`../../../../source/renderer/app/i18n/locales/terms-of-use/${props.locale}.md`);